Introduction
I added an introduction to the paper and related our current output/peak hour bootstrapping to the total compensation of installing solar panels. A lot of articles online discussed whether or not it is worth it for a reader to install solar panels in their area. Figured we could use our output to determine whether it is worth it for Los Angeles residents.